Add driver polling mode
This patch add polling mode by means of hrtimer scheduled work.
This feature could be enabled or disabled through the CONFIG_POLLING_MODE
macro.
This driver modality disable the fifo related sysfs symbols and add the
the polling_rate symbol to controll the scheduling timer. This symbol
take as value the time expressed in msec.

Signed-off-by: Giuseppe Barba <giuseppe.barba@st.com>